refactor: rework with bugfixes and changes for new hilbish version
this commit brings in a few things. i basically
rewrote hilbifetch and it works differently from
how it did before. there is now an api to add items
and order them nicely.

this fixes an issue with calculating ascii length too,
and removes the default config since its done in
the hilbifetch source.